Modification CRUD Client Minimum et Maximum sur RemisePermanente dans Client

+++ Lime/Classes/Client.cs
@@ -33,7 +33,7 @@ namespace Lime
         public string Commentaire { get; set; }
         [Display(Name = "Remise permanente")]
         [Description("Le rabais automatique appliqué, en pourcentage, dont disposera le client.")]
-        public double RemisePermanente { get; set; }
+        public int RemisePermanente { get; set; }
         [Display(Name = "Personne de contact")]
         public string PersonneDeContact { get; set; }
